requirements.txt
click==8.2.1
PyYAML==6.0.2
watchdog==6.0.0
